+++ CHANGELOG.md
@@ -22,6 +22,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **breaking:** `Map::remove` renamed to `Map::try_remove` ([#136])
 - **breaking:** Update to kollect 0.3 ([#138])
 - **breaking:** Add reflected sets ([#140])
+- **fixed:** Fix `<[T; N]>::from_reflect` for `Value::List`. It used to fail but now works
+- **fixed:** `Vec::as_array`, `Vec::as_array_mut`, `Vec::into_array` now
+  correctly return `Some(_)`.
+- **fixed:** `Value::as_array`, `Value::as_array_mut`, `Value::into_array` now
+  correctly return `Some(_)`.
 
 [#90]: https://github.com/EmbarkStudios/mirror-mirror/pull/90
 [#110]: https://github.com/EmbarkStudios/mirror-mirror/pull/110